requirements.txt
aiofiles
aiohttp
m3u8
requests
tqdm
al-utils-almirai
